Ajuste, finalizado upload e adicionado novos campos no cadastro da empresa

+++ application/controllers/assunto_sistema.php
@@ -31,10 +31,10 @@ class Assunto_Sistema extends CI_Controller {
 	
 	public function novo($hel_seqsis_asu) {		
 		$dados = array();
-		$dados['hel_pk_seq_asu']  	= 0;
-		$dados['hel_seqsis_asu']    = base64_decode($hel_seqsis_asu);
-		$dados['hel_titulo_asu']    = '';
-		$dados['hel_link_asu']    	= '';
+		$dados['hel_pk_seq_asu']  = 0;
+		$dados['hel_seqsis_asu']  = base64_decode($hel_seqsis_asu);
+		$dados['hel_titulo_asu']  = '';
+		$dados['hel_link_asu']    = '';
 
 		$dados['ACAO'] = 'Novo';
 
@@ -70,60 +70,46 @@ class Assunto_Sistema extends CI_Controller {
 		global $hel_seqsis_asu;
 		global $hel_titulo_asu;
 		global $hel_link_asu;
-		$config = array();
 		
 		$hel_pk_seq_asu  = $this->input->post('hel_pk_seq_asu');
 		$hel_seqsis_asu  = $this->input->post('hel_seqsis_asu');
-		$hel_titulo_asu  = $this->input->post('hel_titulo_asu');
-		$hel_link_asu  	 = $this->input->post('hel_link_asu');
-		
-		$config['upload_path'] 	 = './uploads/';
-		$config['allowed_types'] = 'gif|jpg|png|pdf|doc|docx';
-		$config['max_size']		 = '1000';
-		
-		echo '<pre>';
-		print_r($config);
-		
-		$this->load->library('upload', $config);
-		
-		if ($this->upload->do_upload('hel_link_asu')){
-			echo 'Upload com sucesso'.$config['upload_path'].$hel_titulo_asu;
-		}else {
-			echo $this->upload->display_errors();
-		}
+		$hel_titulo_asu  = $this->input->post('hel_titulo_asu');	
 
-// 		if ($this->testarDados()) {
-// 			$assunto = array(
-// 				"hel_seqsis_asu"   => $hel_seqsis_asu,
-// 				"hel_titulo_asu"   => $hel_titulo_asu
-// 			);
+		if ($this->testarDados()) {
+			$assunto = array(
+				"hel_seqsis_asu"   => $hel_seqsis_asu,
+				"hel_titulo_asu"   => $hel_titulo_asu,
+				"hel_link_asu"	   => $hel_link_asu
+			);
 			
-// 			if (!$hel_pk_seq_asu) {
-// 				$hel_pk_seq_asu = $this->AssuntoSistemaModel->insert($assunto);
-// 			} else {
-// 				$hel_pk_seq_asu = $this->AssuntoSistemaModel->update($assunto, $hel_pk_seq_asu);
-// 			}
+			if (!$hel_pk_seq_asu) {
+				$hel_pk_seq_asu = $this->AssuntoSistemaModel->insert($assunto);
+			} else {
+				$hel_pk_seq_asu = $this->AssuntoSistemaModel->update($assunto, $hel_pk_seq_asu);
+			}
 
-// 			if (is_numeric($hel_pk_seq_asu)) {
-// 				$this->session->set_flashdata('sucesso', 'Assunto do Sistema salvo com sucesso.');
-// 				redirect('assunto_sistema/index/'.base64_encode($hel_seqsis_asu));
-// 			} else {
-// 				$this->session->set_flashdata('erro', $hel_pk_seq_asu);
-// 				redirect('assunto_sistema/index/'.base64_encode($hel_seqsis_asu));
-// 			}
-// 		} else {
-// 			if (!$hel_pk_seq_asu) {
-// 				redirect('assunto_sistema/novo/'.base64_encode($hel_seqsis_asu));
-// 			} else {
-// 				redirect('assunto_sistema/editar/'.base64_encode($hel_pk_seq_asu).'/'.base64_encode($hel_seqsis_asu));
-// 			}			
-// 		}
+			if (is_numeric($hel_pk_seq_asu)) {
+				$this->session->set_flashdata('sucesso', 'Assunto do Sistema salvo com sucesso.');
+				redirect('assunto_sistema/index/'.base64_encode($hel_seqsis_asu));
+			} else {
+				$this->session->set_flashdata('erro', $hel_pk_seq_asu);
+				redirect('assunto_sistema/index/'.base64_encode($hel_seqsis_asu));
+			}
+		} else {
+			if (!$hel_pk_seq_asu) {
+				redirect('assunto_sistema/novo/'.base64_encode($hel_seqsis_asu));
+			} else {
+				redirect('assunto_sistema/editar/'.base64_encode($hel_pk_seq_asu).'/'.base64_encode($hel_seqsis_asu));
+			}			
+		}
 	}
 	
 	public function apagar($hel_pk_seq_asu, $hel_seqsis_asu) {
-		$res = $this->AssuntoSistemaModel->delete(base64_decode($hel_pk_seq_asu));
-		if ($res) {
-			$this->session->set_flashdata('sucesso', 'Assunto do sistema apagado com sucesso.'); 
+		if ($this->testarApagar(base64_decode($hel_pk_seq_asu))) {
+			$res = $this->AssuntoSistemaModel->delete(base64_decode($hel_pk_seq_asu));
+			if ($res) {
+				$this->session->set_flashdata('sucesso', 'Assunto do sistema apagado com sucesso.'); 
+			}
 		}			
 		redirect('assunto_sistema/index/'.$hel_seqsis_asu);
 	}
@@ -146,6 +132,7 @@ class Assunto_Sistema extends CI_Controller {
 				"hel_codigo_sis"			=> $registro->hel_codigo_sis,
 				"hel_desc_sis"				=> $registro->hel_desc_sis,
 				"hel_titulo_asu"			=> $registro->hel_titulo_asu,
+				"hel_link_asu"				=> base_url()."uploads/assunto_sistema/".$registro->hel_link_asu,
 				"EDITAR_ASSUNTO_SISTEMA"	=> site_url('assunto_sistema/editar/'.base64_encode($registro->hel_pk_seq_asu).'/'.base64_encode($registro->hel_seqsis_asu)),
 				"APAGAR_ASSUNTO_SISTEMA"	=> "abrirConfirmacao('".base64_encode($registro->hel_pk_seq_asu)."','".base64_encode($dados['hel_seqsis_asu'])."')"
 			);
@@ -169,21 +156,84 @@ class Assunto_Sistema extends CI_Controller {
 		global $hel_pk_seq_asu;
 		global $hel_seqsis_asu;
 		global $hel_titulo_asu;
+		global $hel_link_asu;
 
 		$erros    = FALSE;
 		$mensagem = null;
+		
+		$config['upload_path'] 	 = "./uploads/assunto_sistema/";
+		$config['allowed_types'] = 'jpg|jpeg|png|pdf|doc|docx';
+		$config['encrypt_name']  = TRUE;
+		
+		$this->upload->initialize($config);
 
 		$hel_titulo_asu = trim($hel_titulo_asu);
+		
+		if (!empty($hel_pk_seq_asu)){
+			$resultado = $this->AssuntoSistemaModel->get($hel_pk_seq_asu);
+			
+			if ($resultado){
+				$hel_link_asu = $resultado->hel_link_asu;
+			} else {
+				show_error('Não foram encontrados dados do arquivo anterior.', 500, 'Ops, erro encontrado');
+			}	
+		}
+		
 		
 		if (empty($hel_titulo_asu)) {
 			$erros    = TRUE;
 			$mensagem .= "- Titulo não foi preenchido.\n";
 			$this->session->set_flashdata('ERRO_HEL_TITULO_ASU', 'has-error');
 		}
-
+		
+		if (empty($_FILES['hel_link_asu']['size'])){
+			$erros    = TRUE;
+			$mensagem .= "- Selecione um arquivo para upload.\n";
+		}
+		
+		if (!is_dir($config['upload_path'])){
+			$erros    = TRUE;
+			$mensagem .= " - Não é um diretório válido/n";
+		}
+		
+		if (!is_writable($config['upload_path'])){
+			$erros    = TRUE;
+			$mensagem .= " - Não é permitido gravar no diretório/n";
+		}
+		
+		$this->load->library('upload', $config);
+		
+		if (!empty($hel_pk_seq_asu) AND (!empty($_FILES['hel_link_asu']['size']))){
+			if ($resultado){
+				$nome_arquivo = $resultado->hel_link_asu;
+				$caminho = "./uploads/assunto_sistema/";
+				if (file_exists($caminho.$nome_arquivo)){
+					unlink($caminho.$nome_arquivo);
+				}
+			}else {
+				show_error('Não foram encontrados dados do arquivo anterior.', 500, 'Ops, erro encontrado');
+			}	
+		}
+		
+		if (!empty($_FILES['hel_link_asu']['size'])){
+						
+			if ($this->upload->do_upload('hel_link_asu')){
+				$info = $this->upload->data();
+				$hel_link_asu = trim($info['file_name']);
+				if (empty($hel_link_asu)) {
+					$erros    = TRUE;
+					$mensagem .= "- Erro ao gravar o nome do arquivo.\n";
+				}
+			}else {
+				$erros    = TRUE;
+				$mensagem = $this->upload->display_errors();
+			}
+				
+		}
+		
 
 		if ($erros) {
-			$this->session->set_flashdata('titulo_erro', 'Para continuar corrija os seguintes erros:');
+			$this->session->set_flashdata('titulo_erro', 'Para continuar, corrija os seguintes erros:');
 			$this->session->set_flashdata('erro', nl2br($mensagem));
 			
 			$this->session->set_flashdata('ERRO_HEL_ASU', TRUE);
@@ -192,6 +242,34 @@ class Assunto_Sistema extends CI_Controller {
 
 		}
 				
+		return !$erros;
+	}
+	
+	private function testarApagar($hel_pk_seq_asu) {
+		$erros    = FALSE;
+		$mensagem = null;
+	
+		$resultado = $this->AssuntoSistemaModel->get($hel_pk_seq_asu);
+		
+		if ($resultado){
+			$nome_arquivo = $resultado->hel_link_asu;
+			$caminho = "./uploads/assunto_sistema/";
+			if (file_exists($caminho.$nome_arquivo)){
+				unlink($caminho.$nome_arquivo);
+			}else {
+				$erros = TRUE;
+				$mensagem = " - Erro ao verificar se o arquivo existi\n";
+			} 
+		}else {
+			show_error('Não foram encontrados dados do arquivo.', 500, 'Ops, erro encontrado');
+		}
+			
+	
+		if ($erros) {
+			$this->session->set_flashdata('titulo_erro', 'Para apagar corrija os seguintes erros:');
+			$this->session->set_flashdata('erro', nl2br($mensagem));
+		}
+	
 		return !$erros;
 	}
 	
@@ -255,16 +333,6 @@ class Assunto_Sistema extends CI_Controller {
 			$this->session->set_flashdata('erro', nl2br($mensagem));
 			redirect('erro_relatorio');
 		}
-	}
-	
-	public function uploadArquivo($id){
-		
-		$dados = array();
-		
-		$dados['UPLOAD'] 	= site_url('assunto_sistema/uploadArquivo');
-		$dados['CANCELAR'] 	= site_url('assunto_sistema/cancelar');
-		$dados['CANCELAR'] 	= site_url('assunto_sistema/cancelar');
-		$dados['CANCELAR'] 	= site_url('assunto_sistema/excluir');
 	}
 
 }
